+++ osu.Game.Rulesets.Osu/Objects/RepeatPoint.cs
@@ -3,7 +3,6 @@
 
 using System;
 using osu.Game.Beatmaps;
-using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Beatmaps.ControlPoints;
 using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Osu.Judgements;
@@ -11,29 +10,22 @@ using osu.Game.Rulesets.Scoring;
 
 namespace osu.Game.Rulesets.Osu.Objects
 {
-    public class Spinner : OsuHitObject, IHasEndTime
+    public class RepeatPoint : OsuHitObject
     {
-        public double EndTime
-        {
-            get => StartTime + Duration;
-            set => Duration = value - StartTime;
-        }
-
-        public double Duration { get; set; }
-
-        /// <summary>
-        /// Number of spins required to finish the spinner without miss.
-        /// </summary>
-        public int SpinsRequired { get; protected set; } = 1;
+        public int RepeatIndex { get; set; }
+        public double SpanDuration { get; set; }
 
         protected override void ApplyDefaultsToSelf(ControlPointInfo controlPointInfo, BeatmapDifficulty difficulty)
         {
             base.ApplyDefaultsToSelf(controlPointInfo, difficulty);
 
-            SpinsRequired = (int)(Duration / 1000 * BeatmapDifficulty.DifficultyRange(difficulty.OverallDifficulty, 3, 5, 7.5));
+            // Out preempt should be one span early to give the user ample warning.
+            TimePreempt += SpanDuration;
 
-            // spinning doesn't match 1:1 with stable, so let's fudge them easier for the time being.
-            SpinsRequired = (int)Math.Max(1, SpinsRequired * 0.6);
+            // We want to show the first RepeatPoint as the TimePreempt dictates but on short (and possibly fast) sliders
+            // we may need to cut down this time on following RepeatPoints to only show up to two RepeatPoints at any given time.
+            if (RepeatIndex > 0)
+                TimePreempt = Math.Min(SpanDuration * 2, TimePreempt);
         }
 
         public override Judgement CreateJudgement() => new OsuJudgement();