+++ Changelog.md
@@ -182,3 +182,4 @@
 *   合并上游pr
     *   [m] [Fix potentially invalid push in player while already exiting #8355]
     *   [m] [Fix slider ticks/repeats contributing to accuracy #8362]
+    *   [m] [Remove slider implicit judgement #8358]